Moves precompile DoesntRevert tests to its own file
